requirements.txt
# Florentix AI — Render/Railway Backend Dependencies
# This is the ROOT requirements.txt used by the Dockerfile for the TFLite backend
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
tflite-runtime==2.14.0
cachetools==5.3.3
requests==2.31.0
httpx==0.27.0
pydantic==2.6.4
anyio==4.3.0
